Finished refactoring. Crated utils folder and helperFuncs.js
