+++ client/src/utils/ApiServices.js
@@ -1,13 +1,16 @@
 export const addNewCampground = async (campground) => {
   try {
-    const response = await fetch(process.env.REACT_APP_DB_URL + 'campgrounds/', {
-      method: 'POST',
-      headers: { 'Content-type': 'application/json; charset=UTF-8'},
-      body: JSON.stringify(campground)
-    });
+    const response = await fetch(
+      process.env.REACT_APP_DB_URL + 'campgrounds/',
+      {
+        method: 'POST',
+        headers: { 'Content-type': 'application/json; charset=UTF-8' },
+        body: JSON.stringify(campground),
+      }
+    );
     return response.json();
   } catch (error) {
-    console.log('Error from addNewCampround in Services')
+    console.log('Error from addNewCampround in Services');
     alert('Unable to add Campground');
   }
 };
@@ -25,7 +28,9 @@ export const getAllCampgrounds = async () => {
 
 export const getCampgroundById = async (id) => {
   try {
-    const response = await fetch(process.env.REACT_APP_DB_URL + 'campgrounds/' + id);
+    const response = await fetch(
+      process.env.REACT_APP_DB_URL + 'campgrounds/' + id
+    );
     return response.json();
   } catch (err) {
     console.log('Error from getCampgroundById in Services');
